requirements.txt
boto3==1.35.62
botocore==1.35.62
jmespath==1.0.1
mariadb==1.1.11
mysql-connector-python==9.1.0
packaging==24.2
python-dateutil==2.9.0.post0
s3transfer==0.10.3
six==1.16.0
urllib3==2.2.3
PyYAML==6.0.2
